Consolidate error modules
