+++ memory-bank/progress.md
@@ -34,6 +34,7 @@ This file tracks what works, what remains to be built, current status, and known
 - VS Code settings updated for Copilot 1.101+
 - Next.js app scaffolded with Prisma integration
 - SDK Rate Limiter with hourly buckets and 429 back-off logic
+- Consolidated error handling into `src/errors/` directory
 
 <!-- ai:section:whats-left -->
 ## What's Left
