--- src/error/map.ts
@@ -1,5 +1,0 @@
-import { QuestradeErrorPayload } from './types';
-
-export const mapError = (payload: QuestradeErrorPayload): QuestradeErrorPayload => {
-  return payload;
-};
