+++ src/errors/handle.ts
@@ -1,0 +1,9 @@
+import { toQuestradeError } from './toError'
+import { OrderError, QuestradeError, QuestradeErrorPayload } from './types'
+
+export const handleQuestradeError = async <T>(res: Response): Promise<T> => {
+  const err = await toQuestradeError(res)
+  const payload = err.payload as QuestradeErrorPayload
+  if ('orderId' in payload) throw new OrderError(payload)
+  throw err
+}
